Bugfix add option to set content object as current to request
During rendering a fresh content object is created and rendered.
But the extbase configuration manager still uses the current/surrounding
content object. As the TypoScript configuration of mksearch contains setup
that breaks the flexform parsing this option might come in handy if inside
the mksearch plugin some extbase functions like repositories are used.